+++ internal/tempdir/testhelper_test.go
@@ -1,0 +1,20 @@
+package tempdir
+
+import (
+	"os"
+	"testing"
+
+	"gitlab.com/gitlab-org/gitaly/internal/testhelper"
+)
+
+func TestMain(m *testing.M) {
+	os.Exit(testMain(m))
+}
+
+func testMain(m *testing.M) int {
+	defer testhelper.MustHaveNoChildProcess()
+
+	testhelper.ConfigureTestStorage()
+
+	return m.Run()
+}